requirements.txt
grpcio-tools ~= 1.30
matplotlib >= 3.5.0
networkx >= 2.8
plotly >= 5.0.0